Move risc_attribs back to hw/inc
