+++ ttnn/cpp/ttnn/operations/ccl/kernels/edm_fabric/fabric_erisc_datamover_channels.hpp
@@ -11,7 +11,7 @@
 #include "debug/dprint.h"
 #include <tt-metalium/dataflow_api.h>
 #include "tt_metal/hw/inc/ethernet/tunneling.h"
-#include <tt-metalium/risc_attribs.h>
+#include "risc_attribs.h"
 #include "cpp/ttnn/operations/ccl/kernels/edm_fabric/fabric_edm_packet_header.hpp"
 #include "cpp/ttnn/operations/ccl/kernels/edm_fabric/fabric_edm_types.hpp"
 #include "cpp/ttnn/operations/ccl/shared_with_host/hetergeneous_data_structs.hpp"
